[KN]: capturing key strokes events on 'comment' input field
capturing key strokes events on 'comment' input field, checkin to see if
the 'return' was pressed. if it was, im gonna wanna execute a series of
operations, mainly se d the value of the input field in question to the
backend through ajax so that the backend can save the user's comment on
that item concerned, and then, display the comment on the page. make
sure the value of the input field is not an empty string